+++ packages/datetime/src/components/Time/Time.test.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
-import { describe, test, expect, vi } from 'vitest';
+import { describe, test, expect, vi, afterEach } from 'vitest';
 import { userEvent } from '@testing-library/user-event';
+import { TZDate } from '@date-fns/tz';
 
 import { Time } from '@devoinc/genesys-ui-datetime';
 
-import { render, screen } from '@test';
+import { cleanup, render, screen } from '@test';
 
 describe('components', () => {
   describe('Time', () => {
@@ -25,6 +26,25 @@ describe('components', () => {
       await user.type(input, '10');
 
       expect(onChange).toBeCalledTimes(2);
+    });
+
+    test('timezone UTC-06:00', async () => {
+      const onChange = vi.fn();
+      const tz = 'UTC-06:00';
+      render(
+        <Time
+          value={new TZDate(2024, 7, 9, 10, 30, 15, 300, tz)}
+          onChange={onChange}
+          tz={tz}
+        />,
+      );
+      const input = screen.getByLabelText('Time') as HTMLInputElement;
+      expect(input).toBeInTheDocument();
+      expect(input).toHaveDisplayValue('10:30:15');
+    });
+
+    afterEach(() => {
+      cleanup();
     });
   });
 });